+++ metabase_api/__init__.py
@@ -1,1 +1,2 @@
 from .metabase_api import Metabase_API
+from .metabase_api_async import Metabase_API_Async